pull out BaseSaveStateView
